+++ graphspace_python/api/config.py
@@ -1,0 +1,4 @@
+API_HOST = 'www.graphspace.org/api'
+GRAPHS_PATH = '/v1/graphs/'
+LAYOUTS_PATH = '/v1/graphs/{}/layouts/'
+GROUPS_PATH = '/v1/groups/'
