+++ graphspace_python/api/endpoint/groups.py
@@ -1,0 +1,6 @@
+from graphspace_python.api.config import GROUPS_PATH
+
+class Groups(object):
+
+	def __init__(self, client):
+		self.client = client
