+++ graphspace_python/api/errors.py
@@ -1,0 +1,2 @@
+class ErrorHandler(object):
+    pass
